+++ SampleProjects/ExcludeSomething/src/exclude-something.cpp
@@ -1,0 +1,4 @@
+#include "exclude-something.h"
+int excludeSomething(void) {
+  return -1;
+};