+++ SampleProjects/ExcludeSomething/test/null.cpp
@@ -1,0 +1,7 @@
+#include <ArduinoUnitTests.h>
+
+unittest(nothing)
+{
+}
+
+unittest_main()