+++ cpp/arduino/avr/io.h
@@ -96,7 +96,16 @@
 #ifndef _AVR_IO_H_
 #define _AVR_IO_H_
 
-#define _SFR_IO8(io_addr) (io_addr) // this macro is all we need from the sfr file
+#include <stdint.h>
+
+// hardware mocks
+// this set of macros is all we need from the sfr file
+extern volatile uint8_t __ARDUINO_CI_SFR_MOCK[1024];
+#define _SFR_IO8(io_addr) (*(volatile uint8_t *)(__ARDUINO_CI_SFR_MOCK + io_addr))
+#define _SFR_IO16(io_addr) (*(volatile uint16_t *)(__ARDUINO_CI_SFR_MOCK + io_addr))
+#define _SFR_MEM8(io_addr) (*(volatile uint8_t *)(__ARDUINO_CI_SFR_MOCK + io_addr))
+#define _SFR_MEM16(io_addr) (*(volatile uint16_t *)(__ARDUINO_CI_SFR_MOCK + io_addr))
+#define _SFR_MEM32(io_addr) (*(volatile uint32_t *)(__ARDUINO_CI_SFR_MOCK + io_addr))
 
 #if defined (__AVR_AT94K__)
 #  include "ioat94k.h"